+++ packages/node-core/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `ParentProject` to use `untilBlock` as and alias for `block` (#2235)
 
 ### Fixed
-- Historical queries not using the correct block height
+- Historical queries not using the correct block height (#2243)
 
 ## [7.2.0] - 2024-01-30
 ### Changed